+++ src/Controller/MissionController.php
@@ -8,6 +8,7 @@ use App\Repository\MissionRepository;
 use Doctrine\DBAL\Exception\ForeignKeyConstraintViolationException;
 use Knp\Component\Pager\PaginatorInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use Symfony\Component\Config\Definition\Exception\Exception;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
@@ -61,22 +62,28 @@ class MissionController extends AbstractController
         $isedited = false;
         $mission = $missionRepository->find($id);
 
-        $form = $this->createForm(MissionType::class, $mission);
-        $form->handleRequest($request);
-        if ($form->isSubmitted() && $form->isValid())
-        {
+        try {
+            if (!$mission) {
+                throw new Exception('<h2>This mission does not exist, it can\'t be edited</h2>');
+            }
+            $form = $this->createForm(MissionType::class, $mission);
+            $form->handleRequest($request);
+            if ($form->isSubmitted() && $form->isValid())
+            {
+                $entityManager = $this->getDoctrine()->getManager();
 
-            $entityManager = $this->getDoctrine()->getManager();
-
-            $entityManager->persist($mission);
-            $entityManager->flush();
-            $isedited = true;
+                $entityManager->persist($mission);
+                $entityManager->flush();
+                $isedited = true;
+            }
+            return $this->render('mission/edit.html.twig', [
+                'mission' => $mission,
+                'form' => $form->createView(),
+                'isedited' => $isedited
+            ]);
+        } catch (Exception $e) {
+            return new Response($e->getMessage());
         }
-        return $this->render('mission/edit.html.twig', [
-            'mission' => $mission,
-            'form' => $form->createView(),
-            'isedited' => $isedited
-        ]);
     }
 
     /**
@@ -85,14 +92,18 @@ class MissionController extends AbstractController
     public function delete(int $id, MissionRepository $missionRepository): Response
     {
         $mission = $missionRepository->find($id);
-        try{
+        try {
+            if (!$mission) {
+                throw new Exception('<h2>This mission does not exist, it can\'t be deleted</h2>');
+            }
             $entityManager = $this->getDoctrine()->getManager();
             $entityManager->remove($mission);
             $entityManager->flush();
             return new Response('Mission deleted');
         } catch (ForeignKeyConstraintViolationException $e) {
+            return new Response('<h2>Deletion is impossible</h2>');
+        } catch (Exception $e) {
             return new Response($e->getMessage());
         }
-
     }
 }
