+++ src/Controller/SpecialtyController.php
@@ -8,6 +8,7 @@ use App\Repository\SpecialtyRepository;
 use Doctrine\ORM\EntityManagerInterface;
 use Knp\Component\Pager\PaginatorInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use Symfony\Component\Config\Definition\Exception\Exception;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
@@ -60,21 +61,27 @@ class SpecialtyController extends AbstractController
     {
         $isedited = false;
         $spec = $specialtyRepository->find($id);
-
-        $form = $this->createForm(SpecType::class, $spec);
-        $form->handleRequest($request);
-        if ($form->isSubmitted() && $form->isValid())
-        {
-            $entityManager = $this->getDoctrine()->getManager();
-            $entityManager->persist($spec);
-            $entityManager->flush();
-            $isedited = true;
+        try {
+            if (!$spec) {
+                throw new Exception('<h2>This specialty does not exist</h2>');
+            }
+            $form = $this->createForm(SpecType::class, $spec);
+            $form->handleRequest($request);
+            if ($form->isSubmitted() && $form->isValid())
+            {
+                $entityManager = $this->getDoctrine()->getManager();
+                $entityManager->persist($spec);
+                $entityManager->flush();
+                $isedited = true;
+            }
+            return $this->render('specialty/edit.html.twig', [
+                'specialty' => $spec,
+                'form' => $form->createView(),
+                'isedited' => $isedited
+            ]);
+        } catch (Exception $e) {
+            return new Response($e->getMessage());
         }
-        return $this->render('specialty/edit.html.twig', [
-            'specialty' => $spec,
-            'form' => $form->createView(),
-            'isedited' => $isedited
-        ]);
     }
 
     /**
@@ -83,11 +90,17 @@ class SpecialtyController extends AbstractController
     public function delete(int $id, SpecialtyRepository $specialtyRepository): Response
     {
         $spec = $specialtyRepository->find($id);
+        try {
+            if (!$spec) {
+                throw new Exception('<h2>This specialty does not exist</h2>');
+            }
+            $entityManager = $this->getDoctrine()->getManager();
+            $entityManager->remove($spec);
+            $entityManager->flush();
 
-        $entityManager = $this->getDoctrine()->getManager();
-        $entityManager->remove($spec);
-        $entityManager->flush();
-
-        return new Response('Specialty deleted');
+            return new Response('Specialty deleted');
+        } catch (Exception $e) {
+            return new Response($e->getMessage());
+        }
     }
 }
